backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
httpx>=0.27.0
google-generativeai>=0.8.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiofiles>=23.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
